# Reimplement watch cache storage with btree