@wojtek-t 2024-08-27
> I'd really like to see if @wojtek-t can review changes to the watch cache.

It's on my list and I definitely would like to review this PR - just didn't yet get to it.