@serathius 2024-10-25
Got to agreement about the indexer code. PTAL @jpbetz @deads2k @wojtek-t