@dims 2024-10-29
@serathius can we reflect what you mentioned in an earlier comment in the release notes? It currently says `Use btree data structure for watch cache` but should it be `New implementation of watch cache using btree data structure. Implementation is not enabled yet.` (or something equivalent?).

PS: probably need to do the same for the PR title too.