@sftim 2024-10-31
Changelog suggestion
```diff
-New implementation of watch cache using btree data structure. Implementation is not enabled yet.
```
If it's not enabled, I wouldn't changelog it.